Deletion of the createBasicMap function
