+++ app/static/js/traces/_preview.js
@@ -1,12 +1,43 @@
-import { createBasicMap } from "../leaflet/_basic-map.js";
 import { antPath } from "leaflet-ant-path";
+import {
+  addControlGroup,
+  disableControlClickPropagation,
+  getInitialMapState,
+  setMapState,
+} from "../leaflet/_map-utils";
+import { getZoomControl } from "../leaflet/_zoom-control";
+import { getGeolocateControl } from "../leaflet/_geolocate-control";
 
 const tracesDetailsMapContainer = document.querySelector("#trace-map");
-const coords = JSON.parse(tracesDetailsMapContainer.dataset.coords);
 
 if (tracesDetailsMapContainer) {
   console.debug("Initializing trace preview map");
-  const map = createBasicMap(tracesDetailsMapContainer);
+  const coords = JSON.parse(tracesDetailsMapContainer.dataset.coords);
+  const map = L.map(tracesDetailsMapContainer, {
+    zoomControl: false,
+    maxBoundsViscosity: 1,
+    minZoom: 3, // 2 would be better, but is buggy with leaflet animated pan
+    maxBounds: L.latLngBounds(
+      L.latLng(-85, Number.NEGATIVE_INFINITY),
+      L.latLng(85, Number.POSITIVE_INFINITY)
+    ),
+  });
+
+  // Disable Leaflet's attribution prefix
+  map.attributionControl.setPrefix(false);
+
+  // Add native controls
+  map.addControl(L.control.scale());
+
+  // Add controls
+  addControlGroup(map, [getZoomControl(), getGeolocateControl()]);
+
+  // Disable click propagation on controls
+  disableControlClickPropagation(map);
+
+  // TODO: support this on more maps
+  const initialMapState = getInitialMapState(map);
+  setMapState(map, initialMapState, { animate: false });
 
   // Trace path
   const options = {
@@ -20,4 +51,3 @@ if (tracesDetailsMapContainer) {
   const path = antPath(coords, options).addTo(map);
   map.fitBounds(path.getBounds());
 }
-
